requirements.txt
python-dotenv==1.0.1
python-multipart==0.0.20
bs4==0.0.2
fastapi==0.115.4
uvicorn==0.34.0
websockets==13.1
markdown==3.5.2
Pygments==2.17.2  # For additional code highlighting support
SQLAlchemy==2.0.36
langchain-core
langchain-openai
langchain-community
chromadb
langchain-chroma
langchain-anthropic
langgraph
itsdangerous==2.2.0
bcrypt==4.2.1
Jinja2==3.1.5
ollama==0.4.5
psutil==7.0.0
aiosqlite==0.21.0
pydantic==2.11
RestrictedPython==8.0
markitdown[all]